backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
aiofiles>=23.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0

# Media handling
cloudinary>=1.36.0

# HTTP client
httpx>=0.25.0

# Supabase client
supabase>=2.23.0

# AI/ML APIs
google-generativeai>=0.8.0

# Database - PostgreSQL with Supabase (Compatible versions)
sqlalchemy>=1.4.41,<2.1.0
alembic>=1.8.1
psycopg2-binary>=2.9.5
asyncpg>=0.28.0
databases[postgresql]>=0.7.0

# Cache (will add Redis later)
# redis>=4.5.0

# ML libraries (will add when needed)
# pandas>=2.0.0
# scikit-learn>=1.3.0
# numpy>=1.24.0

